update ebpf initialization
